recommendation-engine/requirements.txt
flask==3.1.3
gunicorn==23.0.0
faiss-cpu==1.9.0
numpy==2.1.3
pandas==2.2.3
pyarrow==18.1.0
sentence-transformers==3.3.1
einops==0.8.0
python-dotenv==1.0.1
datasets==4.8.5